spl: cheri: Add Zcheripurecap ABI support for SPL boot
1. Add Zcheripurecap speicfic sections and alignemnts to spl linker
   script.
2. Use cheri_build_infinite_cap() to derive a pointer points to the
   entry of next boot stage from a infinite capability for
   Zcheripurecap ABI.
   Since the m-bit in default infinite capability is in integer pointer
   mode on Zcherihybrid, the next stage will then boot into integer
   pointer mode for Zcheihybrid.
3. Change the variable for storing address from uintptr_t to unsigned
   long.
4. Set the fdt_addr in struct spl_image_info as a capbility pointer
   by using map_physmem() with read/write permission and the size as
   the fdt blob size plus the expanded size for injecting loadables'
   details.

+++ include/spl.h
@@ -256,8 +256,8 @@ enum spl_sandbox_flags {
 struct spl_image_info {
 	const char *name;
 	u8 os;
-	uintptr_t load_addr;
-	uintptr_t entry_point;
+	unsigned long load_addr;
+	unsigned long entry_point;
 #if CONFIG_IS_ENABLED(LOAD_FIT) || CONFIG_IS_ENABLED(LOAD_FIT_FULL)
 	void *fdt_addr;
 #endif
